Move wallet policy serialization logic from account.rs into bip388.rs
The account module should only be concerned with the part of the
(de)serialization that sits above the individual account type.
